Configurable thumbnail size for faces viewer
