+++ tools/manual/cli.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 """ Command Line Arguments for tools """
-from lib.cli import FaceSwapArgs, DirOrFileFullPaths, FileFullPaths
+from lib.cli import FaceSwapArgs, DirOrFileFullPaths, FileFullPaths, Slider
 
 _HELPTEXT = ("This command lets you perform various actions on frames, "
              "faces and alignments files using visual tools.")
@@ -35,4 +35,16 @@ class ManualArgs(FaceSwapArgs):
             "group": "data",
             "help": "Video file or directory containing source frames that faces were extracted "
                     "from."})
+        argument_list.append({
+            "opts": ("-s", "--face-size"),
+            "action": Slider,
+            "type": int,
+            "min_max": (32, 128),
+            "default": 96,
+            "rounding": 16,
+            "dest": "face_size",
+            "help": "Set the thumbnail size for the Faces Viewer. NB: All faces within the video "
+                    "are loaded into RAM, which can take a lot of space. The larger the thumbnail "
+                    "size, the more RAM will be used. The value given here will be scaled to your "
+                    "monitor's DPI. Optional. Default: 96"})
         return argument_list
